Add label selectors for monitors and rules
We introduce a opt-in and opt-out labeling system for prometheusrules
and monitors.

Pod and Service Monitors are opt-out. By setting `monitoring.syn.tools/enabled: "false"`
we can make Prometheus ignore a monitor, even if the namespace it
labeled to be monitored.

PrometheusRules are opt-in. Prometheus will only pick up rules if they
are in a monitored (i.e. labeled) namespace AND the rule is labeled with
`monitoring.syn.tools/enabled: "false"`. This should ensure that rules
are added consciously and prevent a massive import of upstream alerts
that are not actionable or don't meet our standards.